change to Long for version primitives

+++ app/src/main/java/cc/sovellus/vrcaa/api/vrchat/models/FavoriteAvatar.kt
@@ -35,5 +35,5 @@ data class FavoriteAvatar(
     @SerializedName("updated_at")
     var updatedAt: String = "",
     @SerializedName("version")
-    var version: Int = 0
+    var version: Long = 0
 )
